requirements.txt
django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.12
Pillow==10.4.0
djangorestframework-simplejwt==5.3.0

